Tweak gles2-bc to get it to build outside of UIKit

+++ gles2-bc/Sources/OpenGLES/OpenGLESContext.h
@@ -37,8 +37,8 @@ typedef int             GLfixed;
 typedef int             GLclampx;
 
 /* GL types for handling large vertex buffer objects */
-typedef int             GLintptr;
-typedef int             GLsizeiptr;
+typedef signed long int GLintptr;
+typedef signed long int GLsizeiptr;
 
 /* Extensions */
 #define GL_OES_byte_coordinates           1
